+++ NEWS.md
@@ -1,8 +1,8 @@
 # getdata (development version)
 
-- Export `make_xwalk_list()` helper function
-- Reexport functions from `{esri2sf}` and `{sfext}` packages
-- Add new `cache_location_data()` function
+- Export `make_xwalk_list()` helper function (2023-08-14)
+- Reexport functions from `{esri2sf}` and `{sfext}` packages (2023-08-22)
+- Add new `cache_location_data()` function (2023-08-22)
 
 # getdata 0.1.0.9002 (2023-05-02)