+++ NEWS.md
@@ -3,6 +3,7 @@
 - Export `make_xwalk_list()` helper function (2023-08-14)
 - Reexport functions from `{esri2sf}` and `{sfext}` packages (2023-08-22)
 - Add new `cache_location_data()` function (2023-08-22)
+- Add support for named `name` parameter for `get_location_data()` (2023-09-07)
 
 # getdata 0.1.0.9002 (2023-05-02)